+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Governance.hs
@@ -1,16 +1,27 @@
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
 module Cardano.CLI.Shelley.Run.Governance
   ( ShelleyGovernanceCmdError
   , renderShelleyGovernanceError
   , runGovernanceCmd
   ) where
 
-import           Cardano.Prelude
+import           Cardano.Prelude hiding (poll)
 
 import           Data.Aeson
-import qualified Data.ByteString.Lazy as LB
+import           Data.List ((!!))
+import           Data.String(fromString)
 import qualified Data.Text as Text
+import qualified Data.Text.IO as Text
+import qualified Data.Text.Read as Text
+import qualified Data.Text.Encoding as Text
+import qualified Data.ByteString.Char8 as BSC
+import qualified Data.ByteString.Lazy as LB
+import           Formatting (sformat, build)
 
-import           Control.Monad.Trans.Except.Extra (firstExceptT, handleIOExceptT, left, newExceptT)
+import           Control.Monad.Trans.Except.Extra (firstExceptT, handleIOExceptT,
+                   left, newExceptT)
 
 import           Cardano.Api
 import           Cardano.Api.Shelley
@@ -19,9 +30,11 @@ import           Cardano.CLI.Shelley.Key (InputDecodeError, VerificationKeyOrHas
                    readVerificationKeyOrHashOrFile, readVerificationKeyOrHashOrTextEnvFile)
 import           Cardano.CLI.Shelley.Parsers
 import           Cardano.CLI.Types
+import           Cardano.CLI.Shelley.Run.Transaction (ShelleyTxCmdError, readFileTx,
+                   renderShelleyTxCmdError)
 
+import           Cardano.Binary (DecoderError)
 import qualified Cardano.Ledger.Shelley.TxBody as Shelley
-
 
 data ShelleyGovernanceCmdError
   = ShelleyGovernanceCmdTextEnvReadError !(FileError TextEnvelopeError)
@@ -36,7 +49,16 @@ data ShelleyGovernanceCmdError
       !Int
       -- ^ Number of reward amounts
   | ShelleyGovernanceCmdCostModelsJsonDecodeErr !FilePath !Text
-  deriving Show
+  | ShelleyGovernanceCmdUnexpectedKeyType
+      ![TextEnvelopeType]
+      -- ^ Expected key types
+  | ShelleyGovernanceCmdPollOutOfBoundAnswer
+      !Int
+      -- ^ Maximum answer index
+  | ShelleyGovernanceCmdPollInvalidChoice
+  | ShelleyGovernanceCmdTxCmdError !ShelleyTxCmdError
+  | ShelleyGovernanceCmdDecoderError !DecoderError
+  | ShelleyGovernanceCmdVerifyPollError !GovernancePollError
 
 renderShelleyGovernanceError :: ShelleyGovernanceCmdError -> Text
 renderShelleyGovernanceError err =
@@ -56,7 +78,19 @@ renderShelleyGovernanceError err =
       "Error decoding cost model: " <> Text.pack err' <> " at: " <> fp
     ShelleyGovernanceCmdCostModelReadError err' ->
       "Error reading the cost model: " <> Text.pack (displayError err')
-
+    ShelleyGovernanceCmdUnexpectedKeyType expected ->
+      "Unexpected poll key type; expected one of: "
+      <> Text.intercalate ", " (textShow <$> expected)
+    ShelleyGovernanceCmdPollOutOfBoundAnswer nMax ->
+      "Poll answer out of bounds. Choices are between 0 and " <> textShow nMax
+    ShelleyGovernanceCmdPollInvalidChoice ->
+      "Invalid choice. Please choose from the available answers."
+    ShelleyGovernanceCmdTxCmdError txCmdErr ->
+      renderShelleyTxCmdError txCmdErr
+    ShelleyGovernanceCmdDecoderError decoderError ->
+      "Unable to decode metadata: " <> sformat build decoderError
+    ShelleyGovernanceCmdVerifyPollError pollError ->
+      renderGovernancePollError pollError
 
 runGovernanceCmd :: GovernanceCmd -> ExceptT ShelleyGovernanceCmdError IO ()
 runGovernanceCmd (GovernanceMIRPayStakeAddressesCertificate mirpot vKeys rewards out) =
@@ -67,6 +101,12 @@ runGovernanceCmd (GovernanceGenesisKeyDelegationCertificate genVk genDelegVk vrf
   runGovernanceGenesisKeyDelegationCertificate genVk genDelegVk vrfVk out
 runGovernanceCmd (GovernanceUpdateProposal out eNo genVKeys ppUp mCostModelFp) =
   runGovernanceUpdateProposal out eNo genVKeys ppUp mCostModelFp
+runGovernanceCmd (GovernanceCreatePoll prompt choices nonce out) =
+  runGovernanceCreatePoll prompt choices nonce out
+runGovernanceCmd (GovernanceAnswerPoll poll ix) =
+  runGovernanceAnswerPoll poll ix
+runGovernanceCmd (GovernanceVerifyPoll poll metadata) =
+  runGovernanceVerifyPoll poll metadata
 
 runGovernanceMIRCertificatePayStakeAddrs
   :: Shelley.MIRPot
@@ -175,3 +215,124 @@ runGovernanceUpdateProposal (OutputFile upFile) eNo genVerKeyFiles upPprams mCos
     firstExceptT ShelleyGovernanceCmdTextEnvWriteError . newExceptT $
       writeFileTextEnvelope upFile Nothing upProp
 
+runGovernanceCreatePoll
+  :: Text
+  -> [Text]
+  -> Maybe Word
+  -> OutputFile
+  -> ExceptT ShelleyGovernanceCmdError IO ()
+runGovernanceCreatePoll govPollQuestion govPollAnswers govPollNonce (OutputFile out) = do
+  let poll = GovernancePoll{ govPollQuestion, govPollAnswers, govPollNonce }
+
+  let description = fromString $ "An on-chain poll for SPOs: " <> Text.unpack govPollQuestion
+  firstExceptT ShelleyGovernanceCmdTextEnvWriteError . newExceptT $
+    writeFileTextEnvelope out (Just description) poll
+
+  let metadata = asTxMetadata poll
+        & metadataToJson TxMetadataJsonDetailedSchema
+
+  let outPath = out
+        & Text.encodeUtf8 . Text.pack
+
+  liftIO $ do
+    BSC.hPutStrLn stderr $ mconcat
+      [ "Poll created successfully.\n"
+      , "Please submit a transaction using the resulting metadata.\n"
+      ]
+    BSC.hPutStrLn stdout (prettyPrintJSON metadata)
+    BSC.hPutStrLn stderr $ mconcat
+      [ "\n"
+      , "Hint (1): Use '--json-metadata-detailed-schema' and '--metadata-json-file' "
+      , "from the build or build-raw commands.\n"
+      , "Hint (2): You can redirect the standard output of this command to a JSON "
+      , "file to capture metadata.\n\n"
+      , "Note: A serialized version of the poll suitable for sharing with "
+      , "participants has been generated at '" <> outPath <> "'."
+      ]
+
+runGovernanceAnswerPoll
+  :: FilePath
+    -- ^ Poll file
+  -> Maybe Word
+    -- ^ Answer index
+  -> ExceptT ShelleyGovernanceCmdError IO ()
+runGovernanceAnswerPoll pollFile maybeChoice = do
+  poll <- firstExceptT ShelleyGovernanceCmdTextEnvReadError . newExceptT $
+    readFileTextEnvelope AsGovernancePoll pollFile
+
+  choice <- case maybeChoice of
+    Nothing -> do
+      askInteractively poll
+    Just ix -> do
+      validateChoice poll ix
+      liftIO $ BSC.hPutStrLn stderr $ Text.encodeUtf8 $ Text.intercalate "\n"
+        [ govPollQuestion poll
+        , "→ " <> (govPollAnswers poll !! fromIntegral ix)
+        , ""
+        ]
+      pure ix
+
+  let pollAnswer = GovernancePollAnswer
+        { govAnsPoll = hashGovernancePoll poll
+        , govAnsChoice = choice
+        }
+  let metadata =
+        metadataToJson TxMetadataJsonDetailedSchema (asTxMetadata pollAnswer)
+
+  liftIO $ do
+    BSC.hPutStrLn stderr $ mconcat
+      [ "Poll answer created successfully.\n"
+      , "Please submit a transaction using the resulting metadata.\n"
+      ]
+    BSC.hPutStrLn stdout (prettyPrintJSON metadata)
+    BSC.hPutStrLn stderr $ mconcat
+      [ "\n"
+      , "Hint (1): Use '--json-metadata-detailed-schema' and '--metadata-json-file' "
+      , "from the build or build-raw commands.\n"
+      , "Hint (2): You can redirect the standard output of this command to a JSON "
+      , "file to capture metadata."
+      ]
+ where
+  validateChoice :: GovernancePoll -> Word -> ExceptT ShelleyGovernanceCmdError IO ()
+  validateChoice GovernancePoll{govPollAnswers} ix = do
+    let maxAnswerIndex = length govPollAnswers - 1
+    when (fromIntegral ix > maxAnswerIndex) $ left $
+      ShelleyGovernanceCmdPollOutOfBoundAnswer maxAnswerIndex
+
+  askInteractively :: GovernancePoll -> ExceptT ShelleyGovernanceCmdError IO Word
+  askInteractively poll@GovernancePoll{govPollQuestion, govPollAnswers} = do
+    liftIO $ BSC.hPutStrLn stderr $ Text.encodeUtf8 $ Text.intercalate "\n"
+      ( govPollQuestion
+      : [ "[" <> textShow ix <> "] " <> answer
+        | (ix :: Int, answer) <- zip [0..] govPollAnswers
+        ]
+      )
+    liftIO $ BSC.hPutStrLn stderr ""
+    liftIO $ BSC.hPutStr stderr "Please indicate an answer (by index): "
+    txt <- liftIO $ Text.hGetLine stdin
+    liftIO $ BSC.hPutStrLn stderr ""
+    case Text.decimal txt of
+      Right (choice, rest) | Text.null rest ->
+        choice <$ validateChoice poll choice
+      _ ->
+        left ShelleyGovernanceCmdPollInvalidChoice
+
+runGovernanceVerifyPoll
+  :: FilePath
+    -- ^ Poll file
+  -> TxFile
+    -- ^ Signed transaction (answer) file
+  -> ExceptT ShelleyGovernanceCmdError IO ()
+runGovernanceVerifyPoll pollFile (TxFile txFile) = do
+  poll <- firstExceptT ShelleyGovernanceCmdTextEnvReadError . newExceptT $
+    readFileTextEnvelope AsGovernancePoll pollFile
+
+  tx <- firstExceptT ShelleyGovernanceCmdTxCmdError $
+    readFileTx txFile
+
+  signatories <- firstExceptT ShelleyGovernanceCmdVerifyPollError . newExceptT $ pure $
+    verifyPollAnswer poll tx
+
+  liftIO $ do
+    BSC.hPutStrLn stderr "Found valid poll answer, signed by: "
+    BSC.hPutStrLn stdout (prettyPrintJSON signatories)