+++ cardano-api/ChangeLog.md
@@ -4,12 +4,12 @@
 
 ### Features
 
+- New 'Governance.Poll' API implementing [CIP-0094](https://github.com/cardano-foundation/CIPs/pull/496) ([PR 5050](https://github.com/input-output-hk/cardano-node/pull/5050))
 
 ## 1.35.4 -- October 2022
 
 - Expose `Key` interface via `Cardano.Api.Shelley` ([PR4048](https://github.com/input-output-hk/cardano-node/pull/4048))
 - Append, not prepend change output when balancing a transaction ([PR4343](https://github.com/input-output-hk/cardano-node/pull/4343))
-
 - Expose convenience functions `executeQueryCardanoMode`, `determineEra`, `constructBalancedTx` and `queryStateForBalancedTx` ([PR 4446](https://github.com/input-output-hk/cardano-node/pull/4446))
 
 ## 1.35.3 -- August 2022